add and fix
add Reset button in MainWindow
fix deleteFile in all controller

+++ Brew Day!/src/main/instruments/EquipmentController.java
@@ -6,29 +6,29 @@ import java.io.File;
 import java.util.Map;
 
 public class EquipmentController {
-	
+
 	private String filepath;
 	private IOController ioController;
 	private static EquipmentController instance;
-	
-		
+
+
 	private EquipmentController() {
 		super();
 		this.filepath = System.getProperty("user.dir") + "\\src\\Files\\Equipment.txt";
 		this.ioController = new IOController();
 	}
-	
+
 	public static EquipmentController getInstance() {
 		if (instance == null) {
 			instance = new EquipmentController();
 		}
 		return instance;
 	}
-	
+
 	public Equipment extractEquipment() {
 		return (Equipment) ioController.readObjectFromFile(filepath);
 	}
-	
+
 	public void createEquipment(Map<String, Double> instruments) {
 		try {
 			if(instruments == null) throw new NullInstrumentsException();
@@ -39,11 +39,11 @@ public class EquipmentController {
 			System.out.println(e.getMessage());
 		}
 	}
-	
+
 	public void store(Equipment equipment) {
 		ioController.writeObjectToFile(equipment, filepath);
 	}
-	
+
 	public void update(Map<String, Double> instruments) {
 		try {
 			if(instruments == null) throw new NullInstrumentsException();
@@ -54,21 +54,23 @@ public class EquipmentController {
 			System.out.println(e.getMessage());
 		}
 	}
-	
+
 	public void delete(String instrumentName) {
 		Equipment equipment = Equipment.getInstance();
 		equipment.deleteInstrument(instrumentName);
 		store(equipment);
 	}
-	
+
 	//for only testing purpose
 	public void deleteFile() {
 		File file = new File(filepath);
-		
-		if (file.delete()) {
-			System.out.println("\nFile deleted");
-		} else {
-			System.out.println("\nImpossible delete file");
+
+		if (file.exists()) {
+			if (file.delete()) {
+				System.out.println("\nFile deleted");
+			} else {
+				System.out.println("\nImpossible delete file");
+			}
 		}
 	}
 }
