+++ Brew Day!/src/main/resources/StorageController.java
@@ -67,10 +67,12 @@ public class StorageController implements Serializable{
 	//for only testing purpose
 	public void deleteFile() {
 		File file = new File(filepath);
-		if (file.delete()) {
-			System.out.println("\nFile deleted");
-		} else {
-			System.out.println("\nImpossible delete file");
+		if (file.exists()) {
+			if (file.delete()) {
+				System.out.println("\nFile deleted");
+			} else {
+				System.out.println("\nImpossible delete file");
+			}
 		}
 	}
 }
